DEPR: ensure deprecated import still works in inifix 6.0
